requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.1
plotly==5.17.0
seaborn==0.13.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
python-dateutil==2.8.2
pytz==2023.3
supabase==0.4.1  # Supabase Python client